backend/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
ciso8601==2.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-cors-headers==3.10.1
django-polymorphic==3.1.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
drf-nested-routers==0.93.4
drf-yasg==1.20.0
flake8==4.0.1
idna==3.3
inflection==0.5.1
influxdb-client==1.24.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mysqlclient==2.1.0
packaging==21.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==3.2.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
